Increase max width of admin GOV.UK Design System pages
This increases the max width of the pages in the admin interface of
Whitehall in the GOV.UK Design System to 1140px so it is line with
the Bootstrap implementation. This will allow us to develop feature
parity in a like for like way.

In particular we need this to be able to port the history, notes and
fact checking tabs.