Add a feature flag to toggle the use of legacy endpoints
This commit is to test the feature flag in the integration
environment based on user permission. This particular feature flag
will enable toggling the use of legacy endpoints in Asset Manager.

Whitehall calls the /create and /show endpoints in Asset Manager,
using the "legacy_url_path" as an asset identifier.

Co-authored-by: Syed Ali syed.ali@digital.cabinet-office.gov.uk