Republish Worldwide Organisation on Corporate Information Pages changes
Currently, Organisations are republished when a related Corporate Information
Pages changes, but Worldwide Organisations are not. Now that we publish
Worldwide Corporate Information Pages, we want to change this.

This is important because the Worldwide Organisation content item surfaces
information about related Corporate Information Pages on it's `details` rather
than it's `links` (though the links are present as well, they aren't used. This
is consistent with the Organisation content item).