Change attachment update flow to use asset_manager_id if present
All updates triggered by AttachmentUpdater (AccessLimited, DraftStatus etc.)
now check for the existence of assets on the AttachmentData. If assets are present,
subsequent update calls to AssetUpdater are made with each asset's asset_manager_id
(which corresponds to the Asset Manager id). If assets are not present,
the updates default to using the legacy_url_path.

Co-authored-by: Syed Ali syed.ali@digital.cabinet-office.gov.uk