Refactor create worker logic
Saving the asset_manager_id outside the AssetManagerAttachmentSetUploadedToWorker
with the assumption that AttachmentData should be available and there is no
need to wait 0.5s.

Currently running the code only for AttachmentData as to keep the image flow
or other flows that may use carrierwave and use different data models separate
while that logic is written.

Establishing the file version which is required for the asset is currently
done by checking for the presence of the "thumbnail" substring which works
well enough for file attachments.