Introduce use_non_legacy_endpoints flag in attachments controller
The USE_NON_LEGACY_ENDPOINTS permission on the current user is used to set
a flag on the AttachmentData at the controller level.

The AttachmentData instance is then checked in the asset_manager_storage
and if the flag is true we then save an Asset for the corresponding file
the `store!` method has been called with.