Add SocialMediaAccountsController#index in the GOV.UK Design System
I've decided to use a View Component for the summary card as there's
a lot going on here.

If the organisation or worldwide organisation has translations,
each of these use the english translations social media by default.
However, each can be edited to link to a different url.

We've decided to use the optional title for the value of the account,
but it will default to the url if a title has not been provided.

If there are translations each translation has an edit link. If none
are present the edit link sits in the summary card header.